requirements.txt
torch
torchvision
mlflow
dvc
pyyaml